Refactor interface with handlers, create model for password and password\ handler for database

+++ app/src/main/java/ru/nsu/bashev/model/Category.java
@@ -9,7 +9,7 @@ public class Category {
         this.name = name;
     }
 
-    public Category(int id, String name) {
+    public Category(long id, String name) {
         this.id = id;
         this.name = name;
     }
